pyproject.toml
[tool.poetry]
name = "connor-python-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
replit = "^3.0.0"
bs4 = "^0.0.1"
discord = "^1.0.1"
flask = "^1.1.2"
owoify-py = "^1.1.1"
TenGiphPy = "^2.5.0"
google = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
